Stampa Collaboratori: sezione eliminata (disabilitata) Stampa Consulente del lavoro: Rinominato in Stampa Collaboratori; Visualizzazioni dettagliata e sintetica; Riorganizzazione delle colonne; Pulsante stampa nascosto; Nuovo filtro tempo determinato/indeterminato; Nome lavorazione spostato nel riepilogo; Filtro Nominativo spostato come prima voce Scadenzario: Nuovo filtro lavorazione (aggiunta anche colonna in griglia risultati)

+++ VideoSystemWeb/Entity/DatiReportRaw.cs
@@ -28,7 +28,7 @@ namespace VideoSystemWeb.Entity
         private decimal ritenutaAcconto;
         private decimal fattura;
         private decimal fatturaLordo;
-        private int diaria;
+        private decimal diaria;
         private int albergo;
 
         private int tipoPagamento;
@@ -55,7 +55,7 @@ namespace VideoSystemWeb.Entity
         public decimal RitenutaAcconto { get => ritenutaAcconto; set => ritenutaAcconto = value; }
         public decimal Fattura { get => fattura; set => fattura = value; }
         public decimal FatturaLordo { get => fatturaLordo; set => fatturaLordo = value; }
-        public int Diaria { get => diaria; set => diaria = value; }
+        public decimal Diaria { get => diaria; set => diaria = value; }
         public int Albergo { get => albergo; set => albergo = value; }
         public int TipoPagamento { get => tipoPagamento; set => tipoPagamento = value; }
         public string DescrizionePagamento { get => descrizionePagamento; set => descrizionePagamento = value; }
